sdm845-common: configs: initial tweak for the gaming thermal profile
Raise thermal threshold to offer better performance and delay CPU
frequency throttling in games.

According to IEC 60950-1 (2005) standard, electronic devices that are
continuously held in normal use should have a maximum surface temperature
below 55 degree Celsius if the material is metal, and 65 degree
Celsius if the material is glass, porcelain, and vitreous material.

The surface temperature of polaris was measured using an infrared
thermometer with emissivity set to 95. To avoid inaccurate values
from the reflective surface, small pieces of electrical tape were
applied to the device.

According to the infrared thermometer, the warmest area of the device
surface is usually 6-7 degree Celsius lower than fps-therm-adc. When
playing games using the gaming thermal profile, fps-therm-adc can
stabilize around 49-50 degree Celsius, and the warmest surface near
the upper part of the screen is about 43 degree Celsius while other
areas linger between 38 to 42 degree Celsius. Therefore, the result
is well below IEC 60950-1 (2005) standard, and the profile provides
a good balance between an adequate CPU frequency and a comfortable
grip of the device.